Move controller classes to more appropriate package.
