+++ demo/src/main/java/com/novoda/demo/controller/AndroidControllerView.java
@@ -1,4 +1,4 @@
-package com.novoda.demo;
+package com.novoda.demo.controller;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -11,6 +11,7 @@ import android.widget.LinearLayout;
 import android.widget.SeekBar;
 import android.widget.TextView;
 
+import com.novoda.demo.R;
 import com.novoda.notils.exception.DeveloperError;
 
 public class AndroidControllerView extends LinearLayout implements ControllerView {
